+++ build-utils/legacy-player/transcode/src/library/src_manifest/data/shape.rs
@@ -56,7 +56,7 @@ pub enum PathElementKind {
 #[derive(Deserialize, Debug, Clone, Default)]
 pub struct ShapeSettings {
     #[serde(rename="linkToPage")]
-    pub jump_index: Option<usize>,
+    pub jump_index: Option<i32>,
 
     #[serde(rename="toolTipText")]
     pub text: Option<String>,
